requirements.txt
Django~=3.2
psycopg2~=2.8.6
djangorestframework~=3.12.4
coverage~=5.5
python-dotenv~=0.15.0
